pyproject.toml
[tool.poetry]
name = "tapsi-interview"
version = "0.1.0"
description = ""
authors = ["aliraee1 <aliraee1@yahoo.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.7"
django-rest-framework = "^0.1.0"
uvicorn = "^0.30.1"
asgiref = "^3.8.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
